File management and renaming on Adam and Ollie
