--- Assignment/client/yo.txt
@@ -1,1 +1,0 @@
-yo waddup homeslice its NOSE2 assignment
