--- README.md
@@ -1,7 +1,0 @@
-# Nostrils
-yucky boogies
-
-Adam and Gregs cheeky little NOSE2 project
-
-Link to an epic video that basically carried this code:
-https://www.youtube.com/watch?v=qFVoMo6OMsQ&ab_channel=NeuralNine
